+++ Public/Invoke-NUnitForAssembly.ps1
@@ -48,7 +48,7 @@ function Invoke-NUnitForAssembly {
     [string] $DotCoverProcessFilters = '+:nunit-console.exe;+:nunit-console-x86.exe',
     # If set, do not import test results automatically to Teamcity.
     # In this case it is the responsibility of the caller to call 'TeamCity-ImportNUnitReport "$AssemblyPath.$TestResultFilenamePattern.xml"'
-    [switch] $DotNotImportResultsToTeamcity
+    [switch] $DoNotImportResultsToTeamcity
   )
 
   Get-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState -Name 'VerbosePreference'
@@ -99,7 +99,7 @@ function Invoke-NUnitForAssembly {
       Publish-ResultsAndLogs `
         -AssemblyPath $AssemblyPath `
         -TestResultFilenamePattern $TestResultFilenamePattern `
-        -ImportResultsToCIServer (!$DotNotImportResultsToTeamcity.IsPresent)
+        -ImportResultsToCIServer (!$DoNotImportResultsToTeamcity.IsPresent)
   }
 
 }